Done
Results for tests 2 and 3 aren't the same as the ones given. This seems to be because of the fact that I chose the lowest-value term in the case of a tie for a minterm, which Kartik seems to not do. In both tests, though, the number of terms with Xs still are either equal or lower than the results given.